feat: cancel コールバックの実装

+++ src/useGenerateGame.ts
@@ -1,18 +1,24 @@
 // vite では xxx?worker の形でインポートすることで WebWorker として利用できる
 import GenerateGameWorker from './generateGame.worker?worker';
 import { BlockSize } from '@ysk8hori/numberplace-generator';
-import { useEffect, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { MyGame } from './utils/typeUtils';
 import { Difficulty } from './utils/difficulty';
 
-type Result = undefined | { puzzle: MyGame; corrected: MyGame };
+type Result =
+  | {
+      /** ゲームの生成をキャンセルする */
+      cancel: () => void;
+      isGenerating: true;
+    }
+  | { puzzle: MyGame; corrected: MyGame; isGenerating?: false };
 
 /**
  * ナンプレのゲームを生成する hooks
  *
  * @param blockSize ゲームのブロックサイズ
  * @param count ゲームを再生成するたびにインクリメントされる値
- * @returns ゲーム生成結果
+ * @returns ゲーム生成状況や結果
  */
 export default function useGenerateGame({
   blockSize,
@@ -23,19 +29,36 @@ export default function useGenerateGame({
   count: number;
   difficulty: Difficulty;
 }): Result {
-  const [result, setResult] = useState<Result>(undefined);
+  // worker を生成
+  const [worker, setWorker] = useState<Worker | undefined>();
+  // 生成をキャンセルするコールバック
+  const cancel = useCallback(() => {
+    worker?.terminate();
+  }, [count, worker]);
+  const [result, setResult] = useState<Result>({ cancel, isGenerating: true });
+
   useEffect(() => {
-    // worker を生成
-    const worker = new GenerateGameWorker();
-    // ワーカーにゲーム生成を依頼
-    worker.postMessage({ blockSize, difficulty });
+    setWorker(new GenerateGameWorker());
+  }, [blockSize, count]);
+
+  useEffect(() => {
+    if (!worker) return;
     // ワーカーからゲーム生成結果が渡された際の処理を登録
     worker.onmessage = ({ data }) => {
       worker.terminate(); // ワーカーを破棄する
       setResult(data);
     };
-    // return () => worker.terminate(); // ワーカーを破棄する
-  }, [blockSize, count]);
+    // ワーカーにゲーム生成を依頼
+    worker.postMessage({ blockSize, difficulty });
+  }, [count, worker]);
+
+  useEffect(() => {
+    if (result.isGenerating) {
+      // 初回は worker がない状態で cancel コールバックが生成され cancel が機能しないので、
+      // worker ありで cancel が生成され直したら result を更新する。
+      setResult({ cancel, isGenerating: true });
+    }
+  }, [cancel]);
 
   return result;
 }
